feat: persistance buffer logic on motion detection.

+++ src/lit_checker/motion_detection/background_image_processor.py
@@ -30,27 +30,29 @@ class BackgroundImageProcessor:
 
     def __get_background_image_frame(
         self, background_image_path: str, camera_url: str
-    ) -> NDArray[np.int_]:
+    ) -> NDArray[np.uint8]:
         if not os.path.exists(background_image_path):
             background_image = self.__capture_background_image(camera_url, background_image_path)
         else:
             background_image = self.__load_background_image(background_image_path)
         return background_image
 
-    def __load_background_image(self, background_image_path: str) -> NDArray[np.int_]:
+    def __load_background_image(self, background_image_path: str) -> NDArray[np.uint8]:
         cv2_frame = cv2.imread(background_image_path)
-        frame = np.array(cv2_frame, dtype=np.int_)
+        frame = np.array(cv2_frame, dtype=np.uint8)
         self.log.info(f"Loaded background frame from {background_image_path}")
         return frame
 
-    def __capture_background_image(self, url: str, background_image_path: str) -> NDArray[np.int_]:
+    def __capture_background_image(
+        self, url: str, background_image_path: str
+    ) -> NDArray[np.uint8]:
         capture = cv2.VideoCapture(url)
         are_frames, cv2_frame = capture.read()
         if not are_frames:
             self.log.error("Frame could not be captured for background image")
-            frame = np.array([], dtype=np.int_)
+            frame = np.array([], dtype=np.uint8)
         else:
-            frame = np.array(cv2_frame, dtype=np.int_)
+            frame = np.array(cv2_frame, dtype=np.uint8)
             cv2.imwrite(background_image_path, frame)
             self.log.info(f"Wrote background image at {background_image_path}")
         return frame
